refactor: move emulation parameters to separate package
